+++ docker_test/scripts/create.sh
@@ -128,6 +128,12 @@ echo
 
 for IP in $IPLIST; do
   echo "REMOTE_ES_SERVER=\"http://$IP:${REMOTE_PORT}\""
+  if [ "$AUTO_EXPORT" == "y" ]; then
+    # This puts our curatortestenv file where it can be purged easily by destroy.sh
+    cd $SCRIPTPATH
+    cd ..
+    echo "export REMOTE_ES_SERVER=http://$IP:$REMOTE_PORT" > curatortestenv
+  fi
 done
 
 echo